// features/auth/schemas/auth.schema.ts
import { z } from "zod";

export const authLoginSchema = z.object({
  email: z.string().includes("@", { message: "Invalid email" }),
  password: z.string().min(8, "Password must be at least 8 characters long"),
});

export const authSignupSchema = z.object({
  firstName: z.string().min(1, "First name is required"),
  lastName: z.string().min(1, "Last name is required"),
  email: z.string().includes("@", { message: "Invalid email" }),
  password: z.string().min(8, "Password must be at least 8 characters long"),
});